% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_creature.R
\name{create_creature}
\alias{create_creature}
\title{Create an interactive animation in canvas}
\usage{
create_creature(
  interactive = FALSE,
  mouse = FALSE,
  pulse = TRUE,
  color = "darkviolet",
  headradius = 60,
  tickness = 18,
  tentacles = 40,
  friction = 0.02,
  gravity = 0.5,
  wind = -0.5
)
}
\arguments{
\item{interactive}{logical. If TRUE, the animation follows the position of the cursor. Defaults to FALSE}

\item{mouse}{logical. If TRUE, the animation can be moved by the user. Defaults to FALSE}

\item{pulse}{logical, whether the head of the animation should pulsate. Defaults to FALSE}

\item{color}{the color of the animation. Defaults to "darkviolet"}

\item{headradius}{the radius of the head of the animation. Defaults to 60}

\item{tickness}{the thickness of the tentacles of the animation. Defaults to 18}

\item{tentacles}{the animation's number of tentacles. Defaults to 40}

\item{friction}{the value of the friction assigned to the movement of the animation. Defaults to 0.02}

\item{gravity}{the value of gravity assigned to the movement of the animation. Defaults to 0.5}

\item{wind}{A value related to the direction of movement of the animation. Defaults to -0.5.}
}
\value{
an animated form within a canvas
}
\description{
Create an interactive animation in canvas
}
\examples{
if (interactive()) {
  ui <- fluidPage(
    create_creature(
      interactive = FALSE,
      color = "#324C63",
      tentacles = 100,
      tickness = 20,
      friction = 0.3,
      gravity = 0,
      wind = 2
    )
  )

  server <- function(input, output) {

  }


  shinyApp(ui = ui, server = server)
}
}
